VacancyCompetenceScoreRequestDto

+++ atomichunter-rest/src/main/java/net/vniia/skittles/dto/VacancyRespondDto.java
@@ -15,8 +15,8 @@ public class VacancyRespondDto {
     private boolean archive;
     private String email;
     private String fullName;
-    private Integer averageScore;
-    private Integer competenceScoreCount;
+    private long averageScore;
+    private long competenceScoreCount;
     private boolean interviewInviteAccepted;
     private Long interviewId;
 }
